% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lto.R
\name{scale_colour_lto_clinic}
\alias{scale_colour_lto_clinic}
\title{scale_colour_lto_clinic}
\usage{
scale_colour_lto_clinic()
}
\description{
Adds Journal of Clinical Oncology style colours to plot.
}
\examples{
plot + theme_lto() + scale_colour_lto_clinic()
}
