% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lto.R
\name{scale_colour_lto_trans}
\alias{scale_colour_lto_trans}
\title{scale_colour_lto_trans}
\usage{
scale_colour_lto_trans()
}
\description{
Adds Nature Publishing group style colours to plot.
}
\examples{
plot + theme_lto() + scale_colour_lto_trans()
}
