% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lto.R
\name{scale_colour_rel}
\alias{scale_colour_rel}
\title{scale_colour_rel, scale_fill_rel, scale_colour_org and scale_fill_org}
\usage{
scale_colour_rel()
}
\description{
These are personal functions and can be ignored.
}
