% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lto.R
\name{theme_lto}
\alias{theme_lto}
\title{theme_lto}
\usage{
theme_lto(
  base_size = 20,
  base_family = "sans",
  text_size = 20,
  grid = FALSE,
  facet_box = TRUE,
  legend_position = "top",
  legend_title = FALSE
)
}
\arguments{
\item{base_family}{Font of the texts in the plots. Default at 'sans'.}

\item{text_size}{Size of text. Default at 18.}

\item{grid}{Boolean value. If True adds grid to plot. Default at False.}

\item{facet_box}{Boolean value. Draws a box around facets. Only handy when facets are used with facet_wrap. Default at True.}

\item{legend_position}{Defines the position of the legend within the plot. Either choose: 'topright', 'topleft', 'bottomright', 'bottomleft' (all within the figure)or 'top', 'left', 'right', 'bottom' (out of the figure). Default at 'top'.}

\item{legend_title}{Boolean value. If True adds the title of the legend. Set at False.}
}
\value{
Returns the lto theme.
}
\description{
Adds LTO minimal theme layer to ggplots adding + theme_lto to ggplot object. 
To add Journal of Clinical Oncology themed colours add: + scale_fill_lto_clinic() for barplots or scale_colour_lto_clinic() for dotplots, lineplots etc.
To add Nature Publishing Group themed colours add: + scale_fill_lto_trans() for barlots or scale_colour_lto_trans() for dotplots, lineplots etc. 
Any margin warnings can be ignored.
}
\examples{
# Creating an arbitrary dataframe
df <- data.frame(x = c(3,5,7,8,5), y = c(3,7,9,0,3), cline = c('HUB181I', 'HUB181I', 'HUB181I','HUB197', 'HUB197'))
# Creating a ggplot object
plot <- ggplot(data = df, aes(x = x, y = y, col = cline)) +  geom_point(size = 5) 
# Plotting a geom_point with theme_lto
plot + theme_lto()
# Legend to top left.
plot + theme_lto(legend_position = 'topleft')
# With grids.
plot + theme_lto(grid = T)
# With facets but no box around the facets
plot + facet_wrap(~cline) + theme_lto(facet_box = F)
# Adding Nature Publishing group themes colours.
plot + theme_lto() + scale_colour_lto_trans()
}
